updated port/frostfact-api-master/requirements.txt
asgiref==3.8.1
boto3==1.35.41
botocore==1.35.41
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1.2
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-cors-headers==4.5.0
django-filter==24.3
django-grappelli==4.0.1
django-heroku==0.3.1
django-hosts==6.0
django-redis==5.4.0
django-storages==1.14.4
django-subdomain==1.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.15.4
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
Markdown==3.7
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.10
pybase64==1.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
setuptools==70.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.3
whitenoise==6.7.0
